default no print non-critical to stderr; env/config option to control
Ref #9:

Yes I agree with you on this; it's unattractive and too verbose.
I think my original thinking was to match lsusb and ensure that the
user/a script knows that the data is not complete.

Looking into it again, actually lsusb list and tree both don't use the
root_hub descriptors but the usb-ids instead so it does not matter.
Where it will print to stderr, is when using --verbose: 'Couldn't open
device, some information will be missing' for each device that it could
not open as verbose will show string descriptors.

Considering this, I've changed the --lsusb --verbose compatibility mode
to match this exactly and removed the error printing to stderr
otherwise. It can be returned with a
CYME_PRINT_NON_CRITICAL_PROFILER_STDERR env or config key.

capture any profiler error to USBDevice; print in --lsusb --verbose