retry control request on stall
I noticed 'endpoint 0x00: status 0xe0005000 (pipe stalled)' in macOS
`log` nusb profiler. Reading docs, appear this is not unusual and one
should `claim_interface` and `clear_halt`. Seems to resolve; thought the
messages will still be present in `log` as will be on second attempt.

I started digging into this due to Facedancer panicing on device side
with 'usb1.USBErrorTimeout: LIBUSB_ERROR_TIMEOUT [-7]'. With
`--F=libusb` this did not happen. Appears it is with read of Debug
descriptor specifically so have added claim interface to
`ControlRequest` on macOS for this - was not an issue on Linux.

Tools used: `log | ag 'IOUSB'` and Cynthion/Facedancer.