serialize/deserialize blocks inside BlockEditor
Refactoring first go at block editor. Opted to keep serialized blocks in
vec as part of prepare enter editor. The vec can then be manipulated
until committing the changes, when the vec is deserialized back to the
impl Block for the BlockType.

Looked at generics, Box<dyn BlockEnum> but BlockEnum is not dyn
compatiable so did not work. On depand creating the list of blocks each
draw wasn't so nice and think this is a better solution.